LayoutTests/media/av1-codec-per-level-constraints.html

<!DOCTYPE html>
<html>
<head>
    <script src="video-test.js"></script>
    <script>
    function makeVideoConfiguration(width, height, bitrate, framerate) {
        return {
            contentType: 'video/mp4',
            width: width,
            height: height,
            bitrate: bitrate,
            framerate: framerate
        };
    }

    window.addEventListener('load', event => {
        consoleWrite('Level 2.0');
        testExpected(`internals.validateAV1PerLevelConstraints("av01.0.00M.08",
            makeVideoConfiguration(426, 240, 1000, 30))`, true);
        testExpected(`internals.validateAV1PerLevelConstraints("av01.0.00M.08",
            makeVideoConfiguration(2048, 72, 1000, 30))`, true);
        testExpected(`internals.validateAV1PerLevelConstraints("av01.0.00M.08",
            makeVideoConfiguration(128, 1152, 1000, 30))`, true);
        testExpected(`internals.validateAV1PerLevelConstraints("av01.0.00M.08",
            makeVideoConfiguration(640, 480, 1000, 30))`, false);
        testExpected(`internals.validateAV1PerLevelConstraints("av01.0.00M.08",
            makeVideoConfiguration(426, 240, 1000, 60))`, false);
        testExpected(`internals.validateAV1PerLevelConstraints("av01.0.00M.08",
            makeVideoConfiguration(426, 240, 2000000, 30))`, false);
        consoleWrite('Level 2.0 has no High Tier');
        testExpected(`internals.validateAV1PerLevelConstraints("av01.0.00H.08",
            makeVideoConfiguration(426, 240, 1000, 30))`, false);
        consoleWrite('Level 2.1');
        testExpected(`internals.validateAV1PerLevelConstraints("av01.0.01M.08",
            makeVideoConfiguration(640, 360, 1000, 30))`, true);
        testExpected(`internals.validateAV1PerLevelConstraints("av01.0.01M.08",
            makeVideoConfiguration(2816, 99, 1000, 30))`, true);
        testExpected(`internals.validateAV1PerLevelConstraints("av01.0.01M.08",
            makeVideoConfiguration(176, 1584, 1000, 30))`, true);
        testExpected(`internals.validateAV1PerLevelConstraints("av01.0.01M.08",
            makeVideoConfiguration(854, 480, 1000, 30))`, false);
        testExpected(`internals.validateAV1PerLevelConstraints("av01.0.01M.08",
            makeVideoConfiguration(640, 360, 1000, 60))`, false);
        testExpected(`internals.validateAV1PerLevelConstraints("av01.0.01M.08",
            makeVideoConfiguration(640, 360, 4000000, 30))`, false);
        consoleWrite('Level 2.2 is undefined');
        testExpected(`internals.validateAV1PerLevelConstraints("av01.0.02M.08",
            makeVideoConfiguration(640, 360, 1000, 30))`, false);
        consoleWrite('Level 6.1');
        testExpected(`internals.validateAV1PerLevelConstraints("av01.0.17M.08",
            makeVideoConfiguration(7680, 4320, 1000, 60))`, true);
        testExpected(`internals.validateAV1PerLevelConstraints("av01.0.17M.08",
            makeVideoConfiguration(16384, 2176, 1000, 60))`, true);
        testExpected(`internals.validateAV1PerLevelConstraints("av01.0.17M.08",
            makeVideoConfiguration(4096, 8704, 1000, 60))`, true);
        testExpected(`internals.validateAV1PerLevelConstraints("av01.0.17M.08",
            makeVideoConfiguration(15360, 8640, 1000, 60))`, false);
        testExpected(`internals.validateAV1PerLevelConstraints("av01.0.17M.08",
            makeVideoConfiguration(7680, 4320, 200000000, 60))`, false);
        testExpected(`internals.validateAV1PerLevelConstraints("av01.0.17H.08",
            makeVideoConfiguration(7680, 4320, 200000000, 60))`, true);
        testExpected(`internals.validateAV1PerLevelConstraints("av01.0.17H.08",
            makeVideoConfiguration(7680, 4320, 1000000000, 60))`, false);
        consoleWrite('Level 7.0 is undefined')
        testExpected(`internals.validateAV1PerLevelConstraints("av01.0.20M.08",
            makeVideoConfiguration(7680, 4320, 1000, 60))`, false);
        endTest();
    }, { once: true });
    </script>
</head>
<body>
</body>
</html>
